Download win VM outside Vagrant file

+++ release.sh
@@ -38,6 +38,9 @@ fi
 # install dep if not already done
 if [[ ! -d "node_modules" ]]; then
     npm install
+    if [[ $? -ne 0 ]]; then
+        exit 2
+    fi
 fi
 
 echo "Checking that $TAG has been pushed to 'origin'..."
@@ -51,16 +54,17 @@ echo "Remote tag: $REMOTE_TAG"
 echo "Creating the pre-release if it does not exist..."
 ASSETS=`node ./scripts/create-release.js $REMOTE_TAG create`
 
-CESIUM_RELEASE="cesium-$REMOTE_TAG-web"
-if [[ ! -e "$DOWNLOADS/$CESIUM_RELEASE.zip" ]]; then
-    echo "Downloading Cesium web release..."
-    cd $DOWNLOADS
-    wget "https://github.com/duniter/cesium/releases/download/$REMOTE_TAG/$CESIUM_RELEASE.zip"
-    if [[ $? -ne 0 ]]; then
-        exit 2
-    fi
-    cd $ROOT
-fi
+# Downloading web assets (once)
+#CESIUM_RELEASE="cesium-$REMOTE_TAG-web"
+#if [[ ! -f "${DOWNLOADS}/${CESIUM_RELEASE}.zip" ]]; then
+#    echo "Downloading Cesium web release..."
+#    cd $DOWNLOADS
+#    wget "https://github.com/duniter/cesium/releases/download/$REMOTE_TAG/$CESIUM_RELEASE.zip"
+#    if [[ $? -ne 0 ]]; then
+#        exit 2
+#    fi
+#    cd $ROOT
+#fi
 
 
 if [[ "_$EXPECTED_ASSETS" == "_" ]]; then
@@ -85,7 +89,9 @@ for asset in $EXPECTED_ASSETS; do
         echo "Starting Debian build..."
         ./scripts/build.sh make linux $TAG
         DEB_PATH="$PWD/arch/linux/$asset"
-        node ./scripts/upload-release.js $REMOTE_TAG $DEB_PATH
+        if [[ -f "${DEB_PATH}" ]]; then
+          node ./scripts/upload-release.js ${REMOTE_TAG} ${DEB_PATH}
+        fi
       else
         echo "This computer cannot build this asset, required architecture is 'x86_64'. Skipping."
       fi
@@ -97,7 +103,9 @@ for asset in $EXPECTED_ASSETS; do
         echo "Starting Windows build..."
         ./scripts/build.sh make win $TAG
         WIN_PATH="$PWD/arch/windows/$asset"
-        node ./scripts/upload-release.js $REMOTE_TAG $WIN_PATH
+        if [[ -f "${WIN_PATH}" ]]; then
+          node ./scripts/upload-release.js ${REMOTE_TAG} ${WIN_PATH}
+        fi
       else
         echo "This computer cannot build this asset, required architecture is 'x86_64'. Skipping."
       fi
@@ -109,7 +117,9 @@ for asset in $EXPECTED_ASSETS; do
         echo "Starting OSX build..."
         ./scripts/build.sh make win $TAG
         OSX_PATH="$PWD/arch/osx/$asset"
-        node ./scripts/upload-release.js $REMOTE_TAG $OSX_PATH
+        if [[ -f "${OSX_PATH}" ]]; then
+          node ./scripts/upload-release.js ${REMOTE_TAG} ${OSX_PATH}
+        fi
       else
         echo "This computer cannot build this asset, required architecture is 'x86_64'. Skipping."
       fi
@@ -117,9 +127,13 @@ for asset in $EXPECTED_ASSETS; do
   fi
 done
 
-# Clean temporary files
 cd ${ROOT}
-rm ${DOWNLOADS}/cesium-*-web.zip
-rmdir downloads
 
-echo "All the binaries have been uploaded."
+# Clean temporary files
+if [[ $? -eq 0 ]]; then
+  rm ${DOWNLOADS}/cesium-*-web.zip
+  rmdir downloads
+
+  echo "All the binaries have been uploaded."
+fi
+
